pyproject.toml
[tool.poetry]
name = "hipo-drf-exceptions"
version = "0.1.0.3"
description = "An Django app for returning consistent, verbose and easy to parse error messages on Django Rest Framework backends."
authors = ["Hipo <pypi@hipolabs.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/hipo/hipo-drf-exceptions"
repository = "https://github.com/hipo/hipo-drf-exceptions"
keywords = ["python", "django", "django rest framework", "error handling", "exception"]
classifiers = [
    "Framework :: Django",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: Apache Software License",
    'Topic :: Internet :: WWW/HTTP',
    'Topic :: Internet :: WWW/HTTP :: Dynamic Content',
    "Topic :: Software Development :: Libraries :: Python Modules",
    'Programming Language :: Python',
    'Programming Language :: Python :: 2',
    'Programming Language :: Python :: 2.7',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.4',
    'Programming Language :: Python :: 3.5',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
]
include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = ">=2.7, !=3.0, !=3.1, !=3.2, !=3.3, !=3.4, <4"
Django = { version = ">=1.8"}
djangorestframework = { version = ">=3.0"}

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
